list<userdto> is now response type of recommendation endpoint.

+++ App/backend/ScientificPublications/src/main/java/com/sp/ScientificPublications/dto/UserDTO.java
@@ -23,7 +23,7 @@ public class UserDTO {
     @NotEmpty(message = "Provide lastname.")
     private String lastname;
 
-    private Long rank;
+    private Integer rank;
 
     private String token;
